+++ tasmota/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Revert most wifi connectivity changes introduced in 8.1.0.5 (#7746, #7602, #7621)
 - Add initial support for Sensors AHT10 and AHT15 by Martin Wagner (#7596)
+- Add support for Wemos Motor Shield V1 by Denis Sborets (#7764)
 
 ### 8.1.0.8 20200212